Adding baseline functionality for webservice

+++ src/main/java/com/taco/challenge/dto/FoodItem.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class FoodItem {
 
-    private long foodId;
+    private int foodId;
 
     private String foodDescription;
